EV-TTC-main/requirements.txt
Package                   Version
------------------------- --------------
aiohappyeyeballs          2.6.1
aiohttp                   3.11.18
aiosignal                 1.3.2
antlr4-python3-runtime    4.9.3
anyio                     4.9.0
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 3.0.0
async-lru                 2.0.5
async-timeout             5.0.1
attrs                     25.3.0
babel                     2.17.0
beautifulsoup4            4.13.4
bleach                    6.2.0
certifi                   2025.4.26
cffi                      1.17.1
charset-normalizer        3.4.1
colorcet                  3.1.0
comm                      0.2.2
contourpy                 1.3.0
cycler                    0.12.1
datashader                0.17.0
debugpy                   1.8.14
decorator                 4.4.2
defusedxml                0.7.1
exceptiongroup            1.2.2
executing                 2.2.0
fastjsonschema            2.21.1
filelock                  3.18.0
fonttools                 4.57.0
fqdn                      1.5.1
frozenlist                1.6.0
fsspec                    2025.3.2
h11                       0.16.0
h5py                      3.13.0
hdf5plugin                5.1.0
httpcore                  1.0.9
httpx                     0.28.1
hydra-core                1.3.2
idna                      3.10
imageio                   2.37.0
imageio-ffmpeg            0.6.0
importlib_metadata        8.7.0
importlib_resources       6.5.2
ipykernel                 6.29.5
ipython                   8.18.1
ipywidgets                8.1.6
isoduration               20.11.0
jedi                      0.19.2
Jinja2                    3.1.6
joblib                    1.4.2
json5                     0.12.0
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2025.4.1
jupyter                   1.1.1
jupyter_client            8.6.3
jupyter-console           6.6.3
jupyter_core              5.7.2
jupyter-events            0.12.0
jupyter-lsp               2.2.5
jupyter_server            2.15.0
jupyter_server_terminals  0.5.3
jupyterlab                4.4.1
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
jupyterlab_widgets        3.0.14
kiwisolver                1.4.7
lightning                 2.5.1.post0
lightning-utilities       0.14.3
llvmlite                  0.43.0
MarkupSafe                3.0.2
matplotlib                3.9.4
matplotlib-inline         0.1.7
mistune                   3.1.3
moviepy                   1.0.3
mpmath                    1.3.0
multidict                 6.4.3
multipledispatch          1.0.0
narwhals                  1.37.1
nbclient                  0.10.2
nbconvert                 7.16.6
nbformat                  5.10.4
nest-asyncio              1.6.0
networkx                  3.2.1
notebook                  7.4.1
notebook_shim             0.2.4
numba                     0.60.0
numpy                     1.24.0
nvidia-cublas-cu12        12.6.4.1
nvidia-cuda-cupti-cu12    12.6.80
nvidia-cuda-nvrtc-cu12    12.6.77
nvidia-cuda-runtime-cu12  12.6.77
nvidia-cudnn-cu12         9.5.1.17
nvidia-cufft-cu12         11.3.0.4
nvidia-cufile-cu12        1.11.1.6
nvidia-curand-cu12        10.3.7.77
nvidia-cusolver-cu12      11.7.1.2
nvidia-cusparse-cu12      12.5.4.2
nvidia-cusparselt-cu12    0.6.3
nvidia-nccl-cu12          2.26.2
nvidia-nvjitlink-cu12     12.6.85
nvidia-nvtx-cu12          12.6.77
omegaconf                 2.3.0
opencv-python             4.11.0.86
overrides                 7.7.0
packaging                 24.2
pandas                    2.2.3
pandocfilters             1.5.1
param                     2.2.0
parso                     0.8.4
pexpect                   4.9.0
pillow                    11.2.1
pip                       25.0
platformdirs              4.3.7
plotly                    6.0.1
proglog                   0.1.11
projectaria-tools         1.5.8
prometheus_client         0.21.1
prompt_toolkit            3.0.51
propcache                 0.3.1
protobuf                  6.30.2
psutil                    7.0.0
ptyprocess                0.7.0
pure_eval                 0.2.3
pyarrow                   20.0.0
pycparser                 2.22
pyct                      0.5.0
Pygments                  2.19.1
pyparsing                 3.2.3
python-dateutil           2.9.0.post0
python-json-logger        3.3.0
pytorch-lightning         2.5.1.post0
pytz                      2025.2
PyYAML                    6.0.2
pyzmq                     26.4.0
referencing               0.36.2
requests                  2.32.3
rerun-sdk                 0.23.1
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.24.0
scipy                     1.13.1
Send2Trash                1.8.3
setuptools                78.1.1
six                       1.17.0
sniffio                   1.3.1
soupsieve                 2.7
stack-data                0.6.3
sympy                     1.14.0
tensorboardX              2.6.2.2
terminado                 0.18.1
tinycss2                  1.4.0
tomli                     2.2.1
toolz                     1.0.0
torch                     2.7.0
torchaudio                2.7.0
torchmetrics              1.7.1
torchvision               0.22.0
tornado                   6.4.2
tqdm                      4.67.1
traitlets                 5.14.3
triton                    3.3.0
types-python-dateutil     2.9.0.20241206
typing_extensions         4.13.2
tzdata                    2025.2
uri-template              1.3.0
urllib3                   2.4.0
wcwidth                   0.2.13
webcolors                 24.11.1
webencodings              0.5.1
websocket-client          1.8.0
wheel                     0.45.1
widgetsnbextension        4.0.14
xarray                    2024.7.0
yarl                      1.20.0
zipp                      3.21.0
